requirements.txt
aiogram==3.7.0
aiosqlite==0.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.45